Able to just specify new brand for files to be generated

+++ lib/roku_builder/plugins/generator.rb
@@ -5,22 +5,29 @@ require_relative '../../roku_builder_generator'
 
 def default_config()
   return {
-    :component_dir => "brands/core/components",
-    :config_dir => "brands/core/region/US/configs",
+    :component_dir => "brands/%brand%/components",
+    :config_dir => "brands/%brand%/region/US/configs",
     :template_dir => "roku_builder_generator"
   }
 end
 
-def default_output_dir(component_type, base_dir = nil)
+def default_output_dir(component_type, brand = 'core', base_dir = nil)
   base_dir = base_dir || default_config[:component_dir]
+  base_dir = replace_brand(base_dir, brand)
   return "#{base_dir}/#{component_type.capitalize()}s"
 end
 
-def default_config_dir(component_type, base_dir = nil)
+def default_config_dir(component_type, brand = 'core', base_dir = nil)
   base_dir = base_dir || default_config[:config_dir]
-   return "#{base_dir}/#{component_type.capitalize()}s"
+  base_dir = replace_brand(base_dir, brand)
+  return "#{base_dir}/#{component_type.capitalize()}s"
 end
 
+
+def replace_brand(directory, newBrand = 'core')
+  directory["%brand%"]=newBrand
+  return directory
+end
 
 module RokuBuilder
 
@@ -43,20 +50,19 @@ module RokuBuilder
     end
 
 
-    def get_output_dir( component_type, component_name, parent_dir = nil, custom_dir = nil, base_component_dir = nil )
+    def get_output_dir( component_type, component_name, brand, parent_dir = nil, custom_dir = nil, base_component_dir = nil )
       output_dir = get_directory_name(component_type, component_name)
       unless parent_dir.nil? || "" === parent_dir
         output_dir = parent_dir+"/"+output_dir
       end
-      base_dir = custom_dir ? custom_dir : default_output_dir(component_type, base_component_dir)
+      base_dir = !custom_dir.nil? ? replace_brand(custom_dir, brand) : default_output_dir(component_type, brand, base_component_dir)
       return "./#{base_dir}/#{output_dir}"
 
     end
 
-    def get_config_output_dir( component_type, custom_dir = nil, base_config_dir = nil )
-      base_dir = custom_dir ? custom_dir : default_config_dir(component_type, base_config_dir)
+    def get_config_output_dir( component_type, brand, custom_dir = nil )
+      base_dir = !custom_dir.nil? ? replace_brand(custom_dir, brand) : default_config_dir(component_type, brand)
       return "./#{base_dir}"
-
     end
 
     # Hook to add options to the parser
@@ -64,16 +70,21 @@ module RokuBuilder
     #   hash
     def self.parse_options(parser:,  options:)
       parser.separator "Commands:"
+      options[:brand] = 'core'
       parser.on("--generate COMPONENT_TYPE", "Generate a component: manager, module, task, screen") do |component_type|
         options[:generate] = component_type
       end
       parser.on("--name COMPONENT_NAME", "Name of the component") do |component_name|
         options[:name] = component_name
       end
+      parser.separator "Options:"
       parser.on("--extends COMPONENT_NAME", "A component to extend") do |component_name|
        options[:extends] = component_name
       end
-      parser.on("--base-dir", "Base directory for generated code (eg.: '#{default_output_dir('<type>')}')") do |base_dir|
+      parser.on("--for-brand BRAND", "Brand to put component in (default:' core')") do |brand|
+        options[:brand] = brand
+      end
+      parser.on("--base-dir", "Base directory for generated brs/xml code (eg.: '#{default_output_dir('<type>')}')") do |base_dir|
         options[:custom_dir] = base_dir
       end
       parser.on("--with-config", "Add empty config JSON") do |d|
@@ -96,9 +107,11 @@ module RokuBuilder
       #Setup
     end
 
+    def capitalizeFirst(name)
+      name.slice(0,1).capitalize + name.slice(1..-1)
+    end
 
     def get_file_name(component_type, name)
-      name = name.slice(0,1).capitalize + name.slice(1..-1)
       if(component_type == 'screen' && !name.end_with?("Screen"))
         return name+ "Screen"
       end
@@ -109,7 +122,6 @@ module RokuBuilder
     end
 
     def get_directory_name(component_type, name)
-      name = name.slice(0,1).capitalize + name.slice(1..-1)
       if(component_type === 'manager'&& !name.end_with?("Manager"))
         return name+'Manager'
       end
@@ -131,8 +143,9 @@ module RokuBuilder
       config = read_config()
       component_type = options[:generate].downcase
       component_name_parts = options[:name].split('/')
-      component_proper_name = component_name_parts.last
+      component_proper_name = capitalizeFirst(component_name_parts.last)
       component_parent_dir = component_name_parts.first(component_name_parts.size-1).join('/')
+      brand = brand.downcase
 
       component_name = get_file_name(component_type, component_proper_name)
       component = RokuBuilderGenerator::BrsComponent.new(component_name, options[:extends], component_type, @logger)
@@ -140,8 +153,8 @@ module RokuBuilder
       brs_text = component.render("brs")
       json_text = options[:with_config] && component_has_config_json(component_type) ? component.render("json") : nil
 
-      output_dir = get_output_dir(component_type, component_name, component_parent_dir, options[:custom_dir], config[:component_dir])
-      output_config_dir = get_config_output_dir(component_type, options[:config_dir], config[:config_dir])
+      output_dir = get_output_dir(component_type, component_name, brand, component_parent_dir, options[:custom_dir], config[:component_dir])
+      output_config_dir = get_config_output_dir(component_type,  brand, options[:config_dir])
       output_file_name = File.join(output_dir, component_name)
       output_config_file_name = File.join(output_config_dir, component_proper_name)
 
@@ -176,7 +189,11 @@ module RokuBuilder
     def write_file(output_file_name, contents)
       unless contents.nil?
         @logger.unknown output_file_name
-        File.open(output_file_name, "w") { |f| f.write contents }
+        if File.exist?(output_file_name)
+          @logger.warn "#{output_file_name} already exists, skipping"
+        else
+          File.open(output_file_name, "w") { |f| f.write contents }
+        end
       end
     end
